moved header into Header container type element that is stacked on top of timeline.
