+++ src/lib/Timeline.scss
@@ -175,6 +175,21 @@ $weekend: rgba(250, 246, 225, 0.5);
     }
   }
 
+  .rct-sidebar-header {
+    margin: 0;
+    color: $sidebar-color;
+    background: $sidebar-background-color;
+    border-right: $border-width solid $border-color;
+    box-sizing: border-box;
+    border-bottom: $border-width solid $border-color;
+    overflow: hidden;
+  }
+
+  &.rct-sidebar-right {
+    border-right: 0;
+    border-left: $border-width solid $border-color;
+  }
+
   .rct-sidebar {
     overflow: hidden;
     white-space: normal; // was set to nowrap in .rct-outer
@@ -183,21 +198,6 @@ $weekend: rgba(250, 246, 225, 0.5);
     position: relative;
     box-sizing: border-box;
     border-right: $border-width solid $border-color;
-
-    &.rct-sidebar-right {
-      border-right: 0;
-      border-left: $border-width solid $border-color;
-    }
-
-    .rct-sidebar-header {
-      margin: 0;
-      color: $sidebar-color;
-      background: $sidebar-background-color;
-      border-right: $border-width solid $border-color;
-      box-sizing: 'border-box';
-      border-bottom: $border-width solid $border-color;
-      overflow: 'hidden';
-    }
 
     .rct-sidebar-row {
       padding: $list-item-padding;
